requirements.txt
clip==1.0
gradio==4.44.0
Pillow==10.1.0
Pillow==10.4.0
ruclip==0.0.2
torch==2.4.1
